Rename a2notes.tex.txt to a2notes.tex
I uploaded the latex file in case you all want to ever use it as a reference.
I did everything on Ubuntu and it was definitely a PAIN.
You may or may not care but if you do use Debian Linux here is a working template.
You can use $sudo apt-get install texlive-full and that will give you a huge amount of libraries.
To compile it you would then use   $pdfLatex notes.pdf  (do not use pdftex.)
Even though I used the advanced packing tool with texlive it did not recognize any of Dr. Dietz constructs. 
The one I used is as close as i can get it to the way his would look.
So if you are interested and do use Ubuntu then this one may help you. Honestly, for the next project I will probably try to use a third party windows program.